Fix Fast Refresh and simplify Next.js configuration
